+++ CHANGELOG.md
@@ -15,8 +15,10 @@ Changes to this project will be documented in this file.
 - Lots of documentation!
 - Contents to `message_subject.txt` template.
 - Management command, celery tasks for processing published submissions.
+- View and URL for index of mailing list archives.
 ### Changed
 - More fiddling with the badges
+- Message preparation pipeline to reduce template lookups
 ### Removed
 - `Message.subject`